Add files via upload
The Software tool should enable remote-operation of an RF Vector Network Analyzer (VNA) Machine. The measurement data will be then stored in a database. In addition to the measurement data, Meta-data will stored by additional fields that record the conditions of the experiments and other parameter of the experiments.
The projects files:
1.  dbHdf5TablesMod.py:  Module includes the class hdf5DataTable , which uses tables library to store data in a database format and hdf5 file format.
2. vnaGPIBMocMod.py: Uses visa library to connect to GPIB bus and to HP8753C machine. Remote controls the machine and reads samples data from it.
3. rfSampleConsoleMain.py - Main console application that enables user to operate sampling and interfacing with database. Application includes a basic graphic function plotMeas() function

Perliminary version